Move setting files to a new module
