+++ src/routes/settings/get_dns.rs
@@ -10,10 +10,9 @@
 
 use auth::User;
 use config::Env;
-use config_files::SetupVarsEntry;
 use rocket::State;
 use routes::settings::common::as_bool;
-use setup_vars::read_setup_vars;
+use settings::{read_setup_vars, SetupVarsEntry};
 use util::{reply_data, Error, Reply};
 
 /// Get upstream DNS servers
