+++ src/routes/settings/get_network.rs
@@ -10,22 +10,21 @@
 
 use auth::User;
 use config::Env;
-use config_files::SetupVarsEntry::*;
 use hostname::get_hostname;
 use rocket::State;
-use setup_vars::read_setup_vars;
+use settings::{read_setup_vars, SetupVarsEntry};
 use util::{reply_data, Reply};
 
 /// Get Pi-hole local network information
 #[get("/settings/network")]
 pub fn get_network(env: State<Env>, _auth: User) -> Reply {
-    let ipv4_full = read_setup_vars(Ipv4Address, &env)?.unwrap_or_default();
+    let ipv4_full = read_setup_vars(SetupVarsEntry::Ipv4Address, &env)?.unwrap_or_default();
     let ipv4_address: Vec<&str> = ipv4_full.split("/").collect();
-    let ipv6_full = read_setup_vars(Ipv6Address, &env)?.unwrap_or_default();
+    let ipv6_full = read_setup_vars(SetupVarsEntry::Ipv6Address, &env)?.unwrap_or_default();
     let ipv6_address: Vec<&str> = ipv6_full.split("/").collect();
 
     reply_data(json!({
-        "interface": read_setup_vars(PiholeInterface, &env)?.unwrap_or_default(),
+        "interface": read_setup_vars(SetupVarsEntry::PiholeInterface, &env)?.unwrap_or_default(),
         "ipv4_address": ipv4_address[0],
         "ipv6_address": ipv6_address[0],
         "hostname": get_hostname().unwrap_or("unknown".to_owned())
